ISSUE-30: Create operation records page
