+++ package.json
@@ -19,6 +19,7 @@
     "@mui/lab": "^5.0.0-alpha.131",
     "@mui/material": "^5.13.2",
     "@mui/styled-engine": "^5.13.2",
+    "@mui/x-data-grid": "^6.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.11.2"
