+++ src/pages/Users/User.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ContainerLayout from '../../components/ContainerLayout/ContainerLayout';
+import ContainerLayout from '../../components/Layouts/ContainerLayout';
 
 const User: React.FC = () => {
   return (
